refactor(build.yaml): simplify build process and comment out unused caching and build steps
The changes were made to simplify the build process by removing the override flag in the rustup action, changing the way the PATH is set for protoc, and commenting out the caching and build steps that are currently not in use. This is done to make the build process more straightforward and easier to understand, and to avoid unnecessary steps that may slow down the build process.